- Moved from double to floats - Completed audio system with stereo panning and volume based on distance. - Better audio parsing.
Former-commit-id: 11dc674309cc9b9f16484babf08f15c3004b4845

+++ includes/cub3d.h
@@ -120,7 +120,7 @@ typedef struct	s_game
 
 void			initialize_game		(t_game *game, bool screenshot_only);
 void			initialize_game_2	(t_game *game);
-void			player_set_fov				(t_game *game, double fov, bool reset);
+void			player_set_fov				(t_game *game, float fov, bool reset);
 int				game_loop			(t_game *game);
 
 void			parse(int args, char **av, t_game *game, bool *screenshot_only);
